Rename module Cardano.CLI.Run.Legacy* to Cardano.CLI.Legacy.Run*
