+++ cardano-cli/src/Cardano/CLI/Commands/Governance.hs
@@ -8,8 +8,8 @@ import           Cardano.Api
 import           Cardano.Api.Shelley
 
 import           Cardano.Binary (DecoderError)
-import           Cardano.CLI.Run.Legacy.Read (CddlError)
-import           Cardano.CLI.Run.Legacy.StakeAddress
+import           Cardano.CLI.Legacy.Run.Read (CddlError)
+import           Cardano.CLI.Legacy.Run.StakeAddress
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Governance
 import           Cardano.CLI.Types.Key
