+++ cardano-cli/src/Cardano/CLI/EraBased/Run/Governance.hs
@@ -18,7 +18,7 @@ import qualified Cardano.Api.Shelley as Api
 
 import           Cardano.CLI.Commands.Governance
 import           Cardano.CLI.Legacy.Commands.Governance
-import           Cardano.CLI.Run.Legacy.Read (fileOrPipe, readFileTx)
+import           Cardano.CLI.Legacy.Run.Read (fileOrPipe, readFileTx)
 import           Cardano.CLI.Types.Governance
 import qualified Cardano.CLI.Types.Governance as Cli
 import           Cardano.CLI.Types.Key (VerificationKeyOrHashOrFile,
