+++ cardano-cli/src/Cardano/CLI/Legacy/Run/Transaction.hs
@@ -9,7 +9,7 @@
 
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
 
-module Cardano.CLI.Run.Legacy.Transaction
+module Cardano.CLI.Legacy.Run.Transaction
   ( ShelleyTxCmdError(..)
   , renderShelleyTxCmdError
   , runTransactionCmds
@@ -23,9 +23,9 @@ import           Cardano.Api.Shelley
 
 import           Cardano.CLI.EraBased.Legacy
 import           Cardano.CLI.Json.Friendly (friendlyTxBS, friendlyTxBodyBS)
-import           Cardano.CLI.Run.Legacy.Genesis
-import           Cardano.CLI.Run.Legacy.Read
-import           Cardano.CLI.Run.Legacy.Validate
+import           Cardano.CLI.Legacy.Run.Genesis
+import           Cardano.CLI.Legacy.Run.Read
+import           Cardano.CLI.Legacy.Run.Validate
 import           Cardano.CLI.Types.Governance
 import           Cardano.CLI.Types.Legacy
 import           Cardano.CLI.Types.Output
