+++ cardano-cli/test/cardano-cli-test/Test/Cli/ITN.hs
@@ -7,7 +7,7 @@ module Test.Cli.ITN
   , hprop_golden_bech32Decode
   ) where
 
-import           Cardano.CLI.Run.Legacy.Key (decodeBech32)
+import           Cardano.CLI.Legacy.Run.Key (decodeBech32)
 
 import qualified Codec.Binary.Bech32 as Bech32
 import           Control.Monad (void)
