+++ cardano-cli/test/cardano-cli-test/Test/Cli/Shelley/Run/Query.hs
@@ -2,7 +2,7 @@ module Test.Cli.Shelley.Run.Query
   ( hprop_percentage
   ) where
 
-import qualified Cardano.CLI.Run.Legacy.Query as Q
+import qualified Cardano.CLI.Legacy.Run.Query as Q
 import           Cardano.Slotting.Time (RelativeTime (..))
 
 import           Hedgehog (Property, (===))
